refactor: reorganize dateFilterUtils.tx
