+++ superset-frontend/src/explore/components/controls/DateFilterControl/components/AdvancedFrame.tsx
@@ -18,7 +18,7 @@
  */
 import React from 'react';
 import { t } from '@superset-ui/core';
-import { SEPARATOR } from 'src/explore/dateFilterUtils';
+import { SEPARATOR } from 'src/explore/components/controls/DateFilterControl/utils';
 import { Input } from 'src/common/components';
 import { InfoTooltipWithTrigger } from '@superset-ui/chart-controls';
 import { FrameComponentProps } from 'src/explore/components/controls/DateFilterControl/types';
