+++ superset-frontend/src/explore/components/controls/DateFilterControl/utils/index.ts
@@ -18,3 +18,4 @@
  */
 export * from './dateParser';
 export * from './constants';
+export * from './dateFilterUtils';
